Decompose ManagerConsolePage into useSongPrebuffer + useScoring hooks (T7.2)
